build.panel: Build PSID panel data set

Builds a panel data set with id variables pid (unique person identifier) and year from individual PSID family files and supplemental wealth files.

build.panel(
  datadir = NULL,
  fam.vars,
  ind.vars = NULL,
  heads.only = FALSE,
  current.heads.only = FALSE,
  sample = NULL,
  design = "balanced",
  loglevel = INFO
)

Arguments

datadir

either NULL, in which case saves to tmpdir or path to directory containing family files ("FAMyyyy.RData") and individual file ("IND2009ER.RData").

fam.vars

data.frame of variable to retrieve from family files. Can contain see example for required format.

ind.vars

data.frame of variables to get from individual file. In almost all cases this will be the type of survey weights you want to use. don't include id variables ER30001 and ER30002.

heads.only

logical TRUE if user wants household heads only. Household heads in sample year.

current.heads.only

logical TRUE if user wants current household heads only. Distinguishes mover outs heads.

sample

string indicating which sample to select: "SRC" (survey research center), "SEO" (survey for economic opportunity), "immigrant" (immigrant sample), "latino" (Latino family sample). Defaults to NULL, so no subsetting takes place.

design

either character balanced or all or integer. balanced means only individuals who appear in each wave are considered. All means all are taken. An integer value stands for minimum consecutive years of participation, i.e. design=3 means present in at least 3 consecutive waves.

loglevel

one of INFO, WARN and DEBUG. INFO by default.

Value

resulting data.table. the variable pid is the unique person identifier, constructed from ID1968 and pernum

Details

There are several supported approches. Approach one downloads stata data, uses stata to build each wave, then puts it together with `psidR`. The second (recommended) approach downloads all data directly from the psid servers (no Stata needed). For this approach you need to supply the precise names of psid variables - those variable names vary by year. E.g. total family income will have different names in different waves. The function getNamesPSID greatly helps collecting names for all waves.

Merging

The variables interview number in each family file map to the interview number variable of a given year in the individual file. Run example(build.panel) for a demonstration.

Supplements

Notice that support for wealth supplements is disabled! Recent releases of the main family file have wealth data included. Earlier waves must be merged manually, again by variable interview number as above.

Examples

if (FALSE) {
# ################################################
# Real-world example: not run because takes long.
# Build panel with income, wage, age and education
# optionally: add wealth supplements!
# ################################################

# The package is installed with a list of variables
# Alternatively, search for names with \code{\link{getNamesPSID}}
# This is the body of function build.psid()
# (so why not call build.psid() and see what happens!)
r = system.file(package="psidR")
if (small){
  f = fread(file.path(r,"psid-lists","famvars-small.txt"))
  i = fread(file.path(r,"psid-lists","indvars-small.txt"))
} else {
  f = fread(file.path(r,"psid-lists","famvars.txt"))
  i = fread(file.path(r,"psid-lists","indvars.txt"))
}
setkey(i,"name")
setkey(f,"name")
i = dcast(i[,list(year,name,variable)],year~name)
f = dcast(f[,list(year,name,variable)],year~name)
  d = build.panel(datadir="~/datasets/psid/",fam.vars=f,
                 ind.vars=i, 
                 heads.only =TRUE,sample="SRC",
                 design="all")
  save(d,file="~/psid.RData")
}

# ######################################
# reproducible example on artifical data. 
# run this with example(build.panel).
# ######################################

## make reproducible family data sets for 2 years
## variables are: family income (Money) and age

## Data acquisition step:
## run build.panel with sascii=TRUE

# testPSID creates artifical PSID data
td <- testPSID(N=12,N.attr=0)
fam1985 <- data.table::copy(td$famvars1985)
fam1986 <- data.table::copy(td$famvars1986)
IND2019ER <- data.table::copy(td$IND2019ER)

# create a temporary datadir
my.dir <- tempdir()
#save those in the datadir
# notice different R formats admissible
save(fam1985,file=paste0(my.dir,"/FAM1985ER.rda"))
save(fam1986,file=paste0(my.dir,"/FAM1986ER.RData"))
save(IND2019ER,file=paste0(my.dir,"/IND2019ER.RData"))

## end Data acquisition step.

# now define which famvars
famvars <- data.frame(year=c(1985,1986),
                      money=c("Money85","Money86"),
                      age=c("age85","age86"))

# create ind.vars
indvars <- data.frame(year=c(1985,1986),ind.weight=c("ER30497","ER30534"))

# call the builder
# data will contain column "relation.head" holding the relationship code.

d <- build.panel(datadir=my.dir,fam.vars=famvars,
                 ind.vars=indvars,
                 heads.only=FALSE)  
#> INFO [2022-07-14 08:43:20] found FAM1985ER.rda already downloaded
#> INFO [2022-07-14 08:43:20] found FAM1986ER.rda already downloaded
#> INFO [2022-07-14 08:43:20] everything already downloaded. Build dataset now
#> INFO [2022-07-14 08:43:20] psidR: Loading Family data from .rda files
#> INFO [2022-07-14 08:43:20] psidR: loaded individual file: /tmp/Rtmp4u6qsd/IND2019ER.RData
#> INFO [2022-07-14 08:43:20] psidR: total memory load in MB: 1
#> INFO [2022-07-14 08:43:20] psidR: currently working on data for year 1985
#> INFO [2022-07-14 08:43:20] psidR: currently working on data for year 1986
#> INFO [2022-07-14 08:43:20] balanced design reduces sample from 36 to 24
#> INFO [2022-07-14 08:43:20] End of build.panel

# see what happens if we drop non-heads
# only the ones who are heads in BOTH years 
# are present (since design='balanced' by default)
d <- build.panel(datadir=my.dir,fam.vars=famvars,
                 ind.vars=indvars,
                 heads.only=TRUE)  
#> INFO [2022-07-14 08:43:20] found FAM1985ER.rda already downloaded
#> INFO [2022-07-14 08:43:20] found FAM1986ER.rda already downloaded
#> INFO [2022-07-14 08:43:20] everything already downloaded. Build dataset now
#> INFO [2022-07-14 08:43:20] psidR: Loading Family data from .rda files
#> INFO [2022-07-14 08:43:20] psidR: loaded individual file: /tmp/Rtmp4u6qsd/IND2019ER.RData
#> INFO [2022-07-14 08:43:20] psidR: total memory load in MB: 1
#> INFO [2022-07-14 08:43:20] psidR: currently working on data for year 1985
#> INFO [2022-07-14 08:43:20] dropping non-heads leaves 14 obs
#> INFO [2022-07-14 08:43:20] psidR: currently working on data for year 1986
#> INFO [2022-07-14 08:43:20] dropping non-heads leaves 10 obs
#> INFO [2022-07-14 08:43:20] balanced design reduces sample from 17 to 8
#> INFO [2022-07-14 08:43:20] End of build.panel
print(d[order(pid)],nrow=Inf)
#>         money age interview ID1968 pernum sequence relation.head ind.weight
#> 1: 112539.767  60         8   2773     11       20            10 0.06262170
#> 2: 112999.802  61        23   2773     11       20            10 0.14885524
#> 3:   5481.393  26        24   5652      7        1            10 0.16406890
#> 4:   5944.751  27        11   5652      7        1            10 0.07071854
#> 5:  20253.508  39        15   6332     19        1            10 0.05007392
#> 6:  20762.808  40        15   6332     19       20            10 0.61789926
#> 7:  28753.379  47        20   7574      9       20            10 0.98589330
#> 8:  29280.937  48        16   7574      9        1            10 0.16746839
#>        pid year
#> 1: 2773011 1985
#> 2: 2773011 1986
#> 3: 5652007 1985
#> 4: 5652007 1986
#> 5: 6332019 1985
#> 6: 6332019 1986
#> 7: 7574009 1985
#> 8: 7574009 1986

# change sample design to "all": 
# we'll keep individuals if they are head in one year,
# and drop in the other
d <- build.panel(datadir=my.dir,fam.vars=famvars,
                 ind.vars=indvars,heads.only=TRUE,
                 design="all")  
#> INFO [2022-07-14 08:43:20] found FAM1985ER.rda already downloaded
#> INFO [2022-07-14 08:43:20] found FAM1986ER.rda already downloaded
#> INFO [2022-07-14 08:43:20] everything already downloaded. Build dataset now
#> INFO [2022-07-14 08:43:20] psidR: Loading Family data from .rda files
#> INFO [2022-07-14 08:43:20] psidR: loaded individual file: /tmp/Rtmp4u6qsd/IND2019ER.RData
#> INFO [2022-07-14 08:43:20] psidR: total memory load in MB: 1
#> INFO [2022-07-14 08:43:20] psidR: currently working on data for year 1985
#> INFO [2022-07-14 08:43:20] dropping non-heads leaves 14 obs
#> INFO [2022-07-14 08:43:20] psidR: currently working on data for year 1986
#> INFO [2022-07-14 08:43:20] dropping non-heads leaves 10 obs
#> INFO [2022-07-14 08:43:20] End of build.panel
print(d[order(pid)],nrow=Inf)
#>          money age interview ID1968 pernum sequence relation.head ind.weight
#>  1:  13121.005  21        19   2248      5        1            10 0.29350130
#>  2: 112539.767  60         8   2773     11       20            10 0.06262170
#>  3: 112999.802  61        23   2773     11       20            10 0.14885524
#>  4:  12317.814  76         8   3142     24        1            10 0.32997083
#>  5:   9082.691  26        13   3728     15        1            10 0.30773104
#>  6:  43419.757  35        18   3881     13        1            10 0.18911950
#>  7:  12697.017  74        10   5231      3        1            10 0.87676085
#>  8:  44912.794  52        21   5547     21        1            10 0.93575104
#>  9:   5481.393  26        24   5652      7        1            10 0.16406890
#> 10:   5944.751  27        11   5652      7        1            10 0.07071854
#> 11:   9710.608  63         6   6142     20        1            10 0.81243180
#> 12:  20253.508  39        15   6332     19        1            10 0.05007392
#> 13:  20762.808  40        15   6332     19       20            10 0.61789926
#> 14:  29249.320  77        17   7067      6        1            10 0.96214977
#> 15:  28753.379  47        20   7574      9       20            10 0.98589330
#> 16:  29280.937  48        16   7574      9        1            10 0.16746839
#> 17:   3606.359  64         9   7613     22        1            10 0.40040638
#>         pid year
#>  1: 2248005 1986
#>  2: 2773011 1985
#>  3: 2773011 1986
#>  4: 3142024 1986
#>  5: 3728015 1986
#>  6: 3881013 1985
#>  7: 5231003 1986
#>  8: 5547021 1985
#>  9: 5652007 1985
#> 10: 5652007 1986
#> 11: 6142020 1986
#> 12: 6332019 1985
#> 13: 6332019 1986
#> 14: 7067006 1986
#> 15: 7574009 1985
#> 16: 7574009 1986
#> 17: 7613022 1985

file.remove(paste0(my.dir,"/FAM1985ER.rda"),
            paste0(my.dir,"/FAM1986ER.RData"),
            paste0(my.dir,"/IND2019ER.RData"))
#> [1] TRUE TRUE TRUE

# END psidR example

# #####################################################################
# Please go to https://github.com/floswald/psidR for more example usage
# #####################################################################